Convert data.table to a RasterLayer for plotting, etc.

DETAILED DESCRIPTION NEEDED

dt2raster(dt, r, val)

Arguments

dt
data.table object with columns ID, or both X and Y, and the values to assign to the raster specified by column val.

r
Raster* object.

val
The name of the column in dt containing the values for the raster.

Value

A RasterLayer object.

Author

Alex Chubaty

Examples

library(data.table)
library(sp)
library(raster)

#> 
#> Attaching package: ‘raster’

#> The following object is masked from ‘package:data.table’:
#> 
#>     shift

r <- raster(nrows = 10, ncols = 10)
r[] <- 10

# using x,y coordinates
#dt1 <- data.table(X = , Y = , value = r[])

# using pixel ids
dt2 <- data.table(ID = 1L:ncell(r), VALUE = r[])
dt2[, VALUE := sample(1L:10L, ncell(r), replace = TRUE)]

#>       ID VALUE
#>   1:   1     8
#>   2:   2     9
#>   3:   3     6
#>   4:   4     5
#>   5:   5     9
#>   6:   6     3
#>   7:   7     6
#>   8:   8     2
#>   9:   9     6
#>  10:  10     8
#>  11:  11    10
#>  12:  12     8
#>  13:  13    10
#>  14:  14    10
#>  15:  15     2
#>  16:  16     9
#>  17:  17     5
#>  18:  18    10
#>  19:  19     9
#>  20:  20     8
#>  21:  21     2
#>  22:  22     5
#>  23:  23     3
#>  24:  24     4
#>  25:  25     4
#>  26:  26     2
#>  27:  27    10
#>  28:  28     7
#>  29:  29     8
#>  30:  30     6
#>  31:  31     3
#>  32:  32     7
#>  33:  33     8
#>  34:  34     3
#>  35:  35     8
#>  36:  36     7
#>  37:  37     3
#>  38:  38     4
#>  39:  39     2
#>  40:  40    10
#>  41:  41     6
#>  42:  42     9
#>  43:  43     1
#>  44:  44     2
#>  45:  45     1
#>  46:  46     2
#>  47:  47     8
#>  48:  48     6
#>  49:  49     7
#>  50:  50     7
#>  51:  51     1
#>  52:  52     6
#>  53:  53     2
#>  54:  54    10
#>  55:  55     5
#>  56:  56     9
#>  57:  57     9
#>  58:  58     7
#>  59:  59     1
#>  60:  60     7
#>  61:  61     9
#>  62:  62     7
#>  63:  63     3
#>  64:  64     1
#>  65:  65     5
#>  66:  66     8
#>  67:  67     9
#>  68:  68     3
#>  69:  69     6
#>  70:  70     9
#>  71:  71     4
#>  72:  72     9
#>  73:  73     1
#>  74:  74     4
#>  75:  75     3
#>  76:  76     5
#>  77:  77     2
#>  78:  78     2
#>  79:  79     7
#>  80:  80     9
#>  81:  81    10
#>  82:  82     1
#>  83:  83     2
#>  84:  84     1
#>  85:  85     6
#>  86:  86     3
#>  87:  87     4
#>  88:  88     6
#>  89:  89     5
#>  90:  90     2
#>  91:  91     1
#>  92:  92     9
#>  93:  93     6
#>  94:  94     4
#>  95:  95     2
#>  96:  96     6
#>  97:  97     7
#>  98:  98     1
#>  99:  99     5
#> 100: 100     9
#>       ID VALUE

if (interactive())
  plot(dt2raster(dt2, r, "VALUE"))