Determine source of installed packages

Which packages were installed from CRAN, GitHub, Bioconductor, etc.?

pkgSrc(pkg, lib.loc = NULL)

Arguments

pkg
a character string with the package name.

lib.loc
a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. If the default is used, the loaded packages and namespaces are searched before the libraries.

Examples

pkgs <- as.data.frame(installed.packages(), stringsAsFactors = FALSE)
ids <- which(!(pkgs$Priority %in% c("base", "recommended")))
pkgs <- pkgs[ids, ]
pkgs <- pkgs$Package
pkgs[pkgSrc(pkgs) == "CRAN"]

#>   [1] "abind"             "acepack"           "animation"        
#>   [4] "arm"               "askpass"           "assertthat"       
#>   [7] "available"         "beanplot"          "BiocManager"      
#>  [10] "bit"               "bitops"            "blob"             
#>  [13] "brew"              "broom.mixed"       "carData"          
#>  [16] "cellranger"        "checkmate"         "CircStats"        
#>  [19] "classInt"          "clisymbols"        "CodeDepends"      
#>  [22] "colorRamps"        "commonmark"        "cranlike"         
#>  [25] "curl"              "cyclocomp"         "debugme"          
#>  [28] "DEoptim"           "details"           "DiagrammeR"       
#>  [31] "DiagrammeRsvg"     "dichromat"         "downloader"       
#>  [34] "ellipsis"          "evaluate"          "fastdigest"       
#>  [37] "fasterize"         "fastmatch"         "fBasics"          
#>  [40] "flock"             "FNN"               "fpCompare"        
#>  [43] "fs"                "future.callr"      "gdalUtils"        
#>  [46] "gdata"             "geojson"           "geojsonlint"      
#>  [49] "ggimage"           "ggplotify"         "glmm"             
#>  [52] "gmailr"            "gmodels"           "goftest"          
#>  [55] "googledrive"       "gridBase"          "gridExtra"        
#>  [58] "gss"               "gtable"            "gtools"           
#>  [61] "haven"             "highlight"         "highr"            
#>  [64] "httpcode"          "httr"              "igraph"           
#>  [67] "influenceR"        "ini"               "inline"           
#>  [70] "intervals"         "itertools"         "janeaustenr"      
#>  [73] "jpeg"              "jsonify"           "jsonvalidate"     
#>  [76] "kSamples"          "landscapetools"    "later"            
#>  [79] "latticeExtra"      "lazyeval"          "leafem"           
#>  [82] "leaflet.extras"    "leaflet.providers" "leafsync"         
#>  [85] "LearnBayes"        "listenv"           "logging"          
#>  [88] "lwgeom"            "mapdata"           "maps"             
#>  [91] "markdown"          "merTools"          "microbenchmark"   
#>  [94] "miniUI"            "minqa"             "MuMIn"            
#>  [97] "munsell"           "mvtnorm"           "nloptr"           
#> [100] "nortest"           "packrat"           "parsedate"        
#> [103] "permute"           "pkgconfig"         "plogr"            
#> [106] "plotKML"           "plyr"              "png"              
#> [109] "polspline"         "polyclip"          "praise"           
#> [112] "prettycode"        "prettyunits"       "progress"         
#> [115] "proto"             "protolite"         "pryr"             
#> [118] "PtProcess"         "purrr"             "R.cache"          
#> [121] "RandomFields"      "RandomFieldsUtils" "RANN"             
#> [124] "rapidjsonr"        "RApiSerialize"     "raster"           
#> [127] "rbenchmark"        "RColorBrewer"      "rdoc"             
#> [130] "readxl"            "rematch"           "rematch2"         
#> [133] "remotes"           "reshape"           "reshape2"         
#> [136] "rex"               "rhub"              "rJava"            
#> [139] "rmapshaper"        "RODBC"             "roxygen2"         
#> [142] "Rserve"            "rts"               "rvcheck"          
#> [145] "rversions"         "scales"            "sdm"              
#> [148] "selectr"           "sessioninfo"       "shapefiles"       
#> [151] "shinyBS"           "shinycssloaders"   "shinydashboard"   
#> [154] "shinydisconnect"   "showtextdb"        "snakecase"        
#> [157] "snow"              "SnowballC"         "sourcetools"      
#> [160] "spData"            "spdep"             "stabledist"       
#> [163] "stringr"           "SuppDists"         "sylly.en"         
#> [166] "sys"               "tensor"            "TH.data"          
#> [169] "tictoc"            "tidyselect"        "timeDate"         
#> [172] "timeSeries"        "TimeWarp"          "tinytest"         
#> [175] "tokenizers"        "triebeard"         "truncnorm"        
#> [178] "trust"             "udapi"             "urltools"         
#> [181] "uuid"              "versions"          "viridis"          
#> [184] "viridisLite"       "visNetwork"        "webshot"          
#> [187] "whisker"           "whoami"            "widgetframe"      
#> [190] "xlsxjars"          "xml2"              "xopen"            
#> [193] "xtable"            "yaImpute"          "yaml"             
#> [196] "yesno"