Rescale values to a new range

Adapted from https://gis.stackexchange.com/a/194838/47654. This preserves the original distribution of the data.

rescale(x, new.range = c(0, 1), old.range = NULL)

Arguments

x
A numeric vector or Raster* object.

new.range
The lower and upper bounds of the new range. Default c(0,1).

old.range
(optional) The lower and upper bounds of the old range. Default c(min(x), max(x).

Value

A new object whose values have been rescaled.

Examples

rescale(50, old.range = c(0, 100), new.range = c(0, 1)) ## 0.5

#> [1] 0.5

x <- 0:100
rescale(x) # defaults to new range [0,1]

#>   [1] 1.00 0.99 0.98 0.97 0.96 0.95 0.94 0.93 0.92 0.91 0.90 0.89 0.88 0.87 0.86
#>  [16] 0.85 0.84 0.83 0.82 0.81 0.80 0.79 0.78 0.77 0.76 0.75 0.74 0.73 0.72 0.71
#>  [31] 0.70 0.69 0.68 0.67 0.66 0.65 0.64 0.63 0.62 0.61 0.60 0.59 0.58 0.57 0.56
#>  [46] 0.55 0.54 0.53 0.52 0.51 0.50 0.49 0.48 0.47 0.46 0.45 0.44 0.43 0.42 0.41
#>  [61] 0.40 0.39 0.38 0.37 0.36 0.35 0.34 0.33 0.32 0.31 0.30 0.29 0.28 0.27 0.26
#>  [76] 0.25 0.24 0.23 0.22 0.21 0.20 0.19 0.18 0.17 0.16 0.15 0.14 0.13 0.12 0.11
#>  [91] 0.10 0.09 0.08 0.07 0.06 0.05 0.04 0.03 0.02 0.01 0.00

rescale(x, c(-1, 1))

#>   [1]  1.00  0.98  0.96  0.94  0.92  0.90  0.88  0.86  0.84  0.82  0.80  0.78
#>  [13]  0.76  0.74  0.72  0.70  0.68  0.66  0.64  0.62  0.60  0.58  0.56  0.54
#>  [25]  0.52  0.50  0.48  0.46  0.44  0.42  0.40  0.38  0.36  0.34  0.32  0.30
#>  [37]  0.28  0.26  0.24  0.22  0.20  0.18  0.16  0.14  0.12  0.10  0.08  0.06
#>  [49]  0.04  0.02  0.00 -0.02 -0.04 -0.06 -0.08 -0.10 -0.12 -0.14 -0.16 -0.18
#>  [61] -0.20 -0.22 -0.24 -0.26 -0.28 -0.30 -0.32 -0.34 -0.36 -0.38 -0.40 -0.42
#>  [73] -0.44 -0.46 -0.48 -0.50 -0.52 -0.54 -0.56 -0.58 -0.60 -0.62 -0.64 -0.66
#>  [85] -0.68 -0.70 -0.72 -0.74 -0.76 -0.78 -0.80 -0.82 -0.84 -0.86 -0.88 -0.90
#>  [97] -0.92 -0.94 -0.96 -0.98 -1.00

f <- system.file("external/test.grd", package = "raster")
r <- raster::raster(f)
rescale(r) # defaults to new range [0,1]

#> class       : RasterLayer 
#> dimensions  : 115, 80, 9200  (nrow, ncol, ncell)
#> resolution  : 40, 40  (x, y)
#> extent      : 178400, 181600, 329400, 334000  (xmin, xmax, ymin, ymax)
#> coord. ref. : +init=epsg:28992 +towgs84=565.237,50.0087,465.658,-0.406857,0.350733,-1.87035,4.0812 +proj=sterea +lat_0=52.15616055555555 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +ellps=bessel +units=m +no_defs 
#> data source : in memory
#> names       : test 
#> values      : 2.997602e-15, 1  (min, max)
#> 

rescale(r, c(-1, 1))

#> class       : RasterLayer 
#> dimensions  : 115, 80, 9200  (nrow, ncol, ncell)
#> resolution  : 40, 40  (x, y)
#> extent      : 178400, 181600, 329400, 334000  (xmin, xmax, ymin, ymax)
#> coord. ref. : +init=epsg:28992 +towgs84=565.237,50.0087,465.658,-0.406857,0.350733,-1.87035,4.0812 +proj=sterea +lat_0=52.15616055555555 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +ellps=bessel +units=m +no_defs 
#> data source : in memory
#> names       : test 
#> values      : -1, 1  (min, max)
#>